#20: Add static entity file, small fixes
