+++ PLSimulator/entities/entity.py
@@ -1,4 +1,3 @@
-from lib2to3.pytree import Base
 import os
 import pygame
 from pygame import Vector2
@@ -169,43 +168,3 @@ class BaseEntity:
                     collisions.append((this_object, other_object))
 
         return collisions
-
-
-class LandingPad(BaseEntity):
-    '''
-        LandingPad
-    
-        This is the landing pad entity class
-    '''
-
-    def __init__(self):
-        super().__init__(
-            'landing_zone.png',
-            Vector2(256, 16),
-            Vector2(32, 876),  # TODO: change position on reset method
-            Vector2(0, 0),
-            0,
-            100,
-            [],
-            True
-        )
-
-
-class Ground(BaseEntity):
-    '''
-        Ground
-
-        This is the ground environment entity class
-    '''
-
-    def __init__(self):
-        super().__init__(
-            'ground.png',
-            Vector2(640, 16),
-            Vector2(-320, 888),
-            Vector2(0, 0),
-            0,
-            100,
-            [],
-            True
-        )
